requirements.txt
requests
dash
flask
pandas
python-dotenv
dash-bootstrap-components
pyyaml